requirements.txt
attrs==19.1.0
certifi==2019.3.9
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2
django-jet==1.0.8
feedparser==5.2.1
future==0.18.2
gunicorn==20.0.4
idna==2.8
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.6.3
multitasking==0.0.9
nbformat==4.4.0
numpy==1.18.4
pandas==1.0.3
psycopg2==2.7.3.1
pycparser==2.20
pyrsistent==0.14.11
python-dateutil==2.8.1
python-telegram-bot==12.7
pytz==2018.9
requests==2.21.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
tornado==6.0.4
traitlets==4.3.2
Unipath==1.1
urllib3==1.24.1
whitenoise==5.1.0
yfinance==0.1.54
